+++ _admin.php
@@ -26,45 +26,7 @@ $core->auth->setPermissionType('pages',__('manage related pages'));
 
 $core->addBehavior('adminDashboardFavs', array('relatedAdminBehaviors', 'dashboardFavs'));
 $core->addBehavior('adminDashboardFavsIcon', array('relatedAdminBehaviors', 'dashboardFavsIcon'));
-
 $core->addBehavior('sitemapsDefineParts',array('relatedAdminBehaviors','sitemapsDefineParts'));
-$core->addBehavior('initWidgets',array('widgetsRelated','init'));
 
-/**
- *
- */
-class relatedAdminBehaviors
-{
-	/**
-	 *
-	 */
-	public static function sitemapsDefineParts($map) {
-		$map[__('Related pages')] = 'related';
-	}
 
-	public static function dashboardFavs($core,$favs) {
-		$favs['related'] = new ArrayObject(array(
-			'related',
-			__('Related pages'),
-			'plugin.php?p=related',
-			'index.php?pf=related/imgs/icon.png',
-			'index.php?pf=related/imgs/icon-big.png',
-			'usage,contentadmin',
-			null,
-			null));
-	}
-
-    public static function dashboardFavsIcon($core, $name, $icon) {
-        if ($name == 'related') {
-            $params = new ArrayObject();
-            $params['post_type'] = 'related';
-            $page_count = $core->blog->getPosts($params,true)->f(0);
-            if ($page_count > 0) {
-                $str_pages = ($page_count > 1) ? __('%d related pages') : __('%d related page');
-                $icon[0] = sprintf($str_pages,$page_count);
-            } else {
-                $icon[0] = __('Related pages');
-            }
-        }
-    }
-}
+require dirname(__FILE__).'/_widgets.php';